[refactor] api응답 메시지 수정 및 dto 경로 변경
