+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/application/service/CompanyClient.java
@@ -2,7 +2,7 @@ package com.shoonglogitics.orderservice.domain.delivery.application.service;
 
 import java.util.UUID;
 
-import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryCompanyInfo;
+import com.shoonglogitics.orderservice.domain.delivery.application.service.dto.CreateDeliveryCompanyInfo;
 import com.shoonglogitics.orderservice.global.common.vo.UserRoleType;
 
 public interface CompanyClient {
