+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/application/service/HubClient.java
@@ -3,7 +3,7 @@ package com.shoonglogitics.orderservice.domain.delivery.application.service;
 import java.util.List;
 import java.util.UUID;
 
-import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryRoutesInfo;
+import com.shoonglogitics.orderservice.domain.delivery.application.service.dto.CreateDeliveryRoutesInfo;
 import com.shoonglogitics.orderservice.global.common.vo.UserRoleType;
 
 public interface HubClient {
