+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/application/service/OrderClient.java
@@ -2,7 +2,7 @@ package com.shoonglogitics.orderservice.domain.delivery.application.service;
 
 import java.util.UUID;
 
-import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryOrderInfo;
+import com.shoonglogitics.orderservice.domain.delivery.application.service.dto.CreateDeliveryOrderInfo;
 import com.shoonglogitics.orderservice.global.common.vo.UserRoleType;
 
 public interface OrderClient {
