+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/infrastructure/external/client/impl/HubClientImpl.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Component;
 
-import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryRoutesInfo;
 import com.shoonglogitics.orderservice.domain.delivery.application.service.HubClient;
+import com.shoonglogitics.orderservice.domain.delivery.application.service.dto.CreateDeliveryRoutesInfo;
 import com.shoonglogitics.orderservice.domain.delivery.infrastructure.external.client.feign.HubFeignClient;
 import com.shoonglogitics.orderservice.domain.delivery.infrastructure.external.mapper.HubMapper;
 import com.shoonglogitics.orderservice.global.common.vo.UserRoleType;
