+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/infrastructure/external/mapper/OrderMapper.java
@@ -1,6 +1,6 @@
 package com.shoonglogitics.orderservice.domain.delivery.infrastructure.external.mapper;
 
-import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryOrderInfo;
+import com.shoonglogitics.orderservice.domain.delivery.application.service.dto.CreateDeliveryOrderInfo;
 import com.shoonglogitics.orderservice.domain.delivery.infrastructure.external.dto.FeignOrderResponse;
 
 public final class OrderMapper {
