+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/presentation/dto/CreateDeliveryResponse.java
@@ -5,13 +5,11 @@ import java.util.UUID;
 import com.shoonglogitics.orderservice.domain.delivery.application.dto.CreateDeliveryResult;
 
 public record CreateDeliveryResponse(
-	UUID deliveryId,
-	String message
+	UUID deliveryId
 ) {
 	public static CreateDeliveryResponse from(CreateDeliveryResult result) {
 		return new CreateDeliveryResponse(
-			result.deliveryId(),
-			result.maeesage()
+			result.deliveryId()
 		);
 	}
 }
